Support completely empty test classes
In the case of an empty test class similar to:

class SomeTest < Test::Unit::TestCase
end

`m` was bombing out because `exit!` expects a non-nil value. The
`Runner#run` method was only returning values in the case of tests that
were found or tests not found within a file that does have tests. The
third case of no tests at all was not being covered and would result in
a `nil` being returned.