Extract a strategy for each supported framework into a proper file
This is a refactor that should not affect any public facing methods.
Handling different frameworks (or different versions of the same
framework) was becoming messy really fast. The conditionals for each of
them were scattered through the code. Thus, the solution was to use the
strategy pattern (almost), and abstract the differences into what I
called a runner.

The different runners could not be injected into the initializer as a
dependency, because at that point the testing framework is not yet
loaded, which means there is no way of knowing what runner to use. So,
it is lazy loaded, and only calculated when it is first used.

Removed the main Runner class from the M module file, in order to make it
simpler and easier to understand and reason about.

The framework runner class now uses instance methods instead of class
methods, mostly because of a personal preference, but also because I
feel it instance methods are more flexible, and easier to change.