Rescue RuntimeError with exit!(), not exit()
WHY:
In parser.rb, RunTime error is being rescued with exit(). By default
this returns a status code of 0:

http://ruby-doc.org/core-2.3.1/Kernel.html#method-i-exit

CI builds depend on status codes for each step in the build process to
determine whether the build passed or failed. Since exit() is being used
here, test suites that experience runtime errors will still exit with a
0 status code.

By contrast, exit!() defaults to returning a status code of 1 (failure):

http://ruby-doc.org/core-2.3.1/Kernel.html#method-i-exit-21

This is the desired behavior for handling a runtime error during tests.