Work correctly on Ruby 4.0
On Ruby 4.0, method.source_location returns an array with 5 entries
instead of 2.  Additional entries are start column, end line, and
end column.

Don't assume start_line is the last entry in the array, always use
offset 1.

If the end line is available, use it instead of splitting the method
source and trying to determine the number of lines in it.

Do not require method_source on Ruby 4.0+, since it will not be used.