Add mutex_m and logger to the CI gemfiles depending on Ruby version
mutex_m starts warning in Ruby 3.3, and fails in Ruby 3.4.
logger starts warning in Ruby 3.4, and fails in Ruby 4.0.

However, one of the warnings in causes a test failure, as m is
testing for specific output. Adding mutex_m and logger to the CI
gemfiles should work around the issue.

Note that this is only an issue because the tests use an old
version of active_support.