Revert getting the end_line location from the source location
This was reverted in Ruby 4.0, with only about a week until the
release. While only the require_relative line actually needed a
change, since the code itself would be correct even with the
revert, I think it's best to use simpler code.

It's likely Ruby 4.1 will include support for this, so comment out
the code to make it easy to change back.